fix(application): declare every replacement document on task supersede
The replacement task's create mutation carried no storage context, so the
registry located the package directory itself. That bare prefix path is never
covered by the exact portable-path scopes minted from the observed write set,
and admission rejected every supersede with TOKEN_PATH_SCOPE_DENIED.

Split the task disposition compilers and their shared compilation primitives
into their own modules so the file stays within the complexity budget.
